.xlsx: what LibreOffice writes reads right, and what office42 writes reopens
Reading:

- LibreOffice spells its booleans out -- showGridLines="true",
  locked="true", sheet="true" -- and the reader took them with atoi,
  which makes "true" 0: gridlines and zeros went off, protection went,
  and saving wrote showZeros="0" for Excel to hide the zeros.  Every
  integer attribute now takes true and false, and is held well inside
  an int so that the arithmetic after it cannot wrap.
- Gnumeric writes a border's sides as start and end, which were
  ignored; they are left and right.
- A rich run's size was cast to eight bits, so 64pt came back as 0.
- A form control's list on another sheet lost its sheet and read this
  one's cells.
- A contour chart went out as surface3DChart and came back a surface;
  it is Excel's surfaceChart, both ways.

Against hostile files: a zip whose directory names the same bytes many
times inflated them each time (34 KB became a gigabyte); a part is read
once and the whole archive may come to 200 times its size.  A pivot's
aggregate is one the table has.  A merge larger than a whole column,
or a What-If table larger than 16,384 cells, costs minutes to lay out;
the merge is dropped and the table keeps the values the file has.

Writing:

- Text before a rich cell's first run was left out: "Hello World" with
  World bold was saved as "World".
- A formula the parser could not read -- [1]Sheet1!$A$1*2, a link to
  another book -- was saved as #NAME?; it goes out as it was written.
- A font name, a chart's axis format and a control's list range were
  written without escaping, so AT&T Sans or "$"#,##0 made a part that
  was not XML and office42 could not open its own file.
- Print_Area and Print_Titles put a sheet's name in quotes without
  doubling a quote inside it, so LibreOffice found no print area on
  "Bob's Data".  They are quoted as a formula quotes them.
- A 3-D line chart had a marker and two axes where the schema has no
  marker and three, stacked 3-D columns had an overlap, surfaces and
  contours had no series axis, and a shape with no words had a text
  body with no paragraph in it.
- The window let GTK set LC_NUMERIC from the machine as it started, so
  on a German desktop numbers went out as defaultColWidth="10,7143";
  numbers are in C again once GTK is up, as main() meant them to be.

Checked: LibreOffice's .xlsx of a sheet reads with gridlines and zeros
on.  All seventeen chart kinds, 2-D and 3-D, with the axis format
"$"#,##0, validate against ISO 29500's dml-chart.xsd with xmllint, and
the contour reloads as a contour.  "Hello <b>World</b>" from a .fods,
a run in AT&T Sans and =[1]Sheet1!$A$1*2 survive a save and reload as
.xlsx.  The window under xvfb with LC_ALL=de_DE.UTF-8 writes
defaultColWidth="10.7143".  The reviewer's bomb32.xlsx peaks at 53 MB
(1,045 MB before), dt1m, dt4000 and merge8000 load in 0.03 s, and
pv-agg7, pv-aggbig and dupname load and refresh clean under ASan.

+++ src/io/o42-gnumeric.c
@@ -3247,7 +3247,7 @@ placed:
               run.fmt.italic = (flags & 2) != 0;
               run.fmt.underline = (flags & 4) != 0;
               run.fmt.strikeout = (flags & 8) != 0;
-              run.fmt.size = (guint8) CLAMP (atoi (fields[2]), 2, 800);
+              run.fmt.size = CLAMP (atoi (fields[2]), 2, 800);
               run.fmt.colour = (guint32) g_ascii_strtoull (fields[3], NULL, 16);
               run.fmt.family = g_intern_string (fields[4]);
               g_array_append_val (runs, run);
